requirements.txt
streamlit>=1.36
streamlit-autorefresh>=1.0.1
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
requests>=2.32
python-binance>=1.0.19
openai>=1.40.0
google-generativeai>=0.7.2
pytest>=8.3
